+++ R/zzz.R
@@ -1,0 +1,26 @@
+.pkgenv <- new.env(parent = emptyenv())
+
+.onAttach <- function(...){
+  ## Retrieve Year Information
+  date <- date()
+  x <- regexpr("[0-9]{4}", date)
+  this.year <- substr(date, x[1], x[1] + attr(x, "match.length") - 1)
+  
+  # Retrieve Current Version
+  this.version = packageVersion("Riemann")
+  
+  ## Print on Screen
+  packageStartupMessage("** ------------------------------------------------------- **")
+  packageStartupMessage("**  Riemann || Learning with Data on Riemannian Manifolds      ")
+  packageStartupMessage("**")
+  packageStartupMessage("** Version    : ",this.version,"       (",this.year,")",sep="")
+  packageStartupMessage("** Maintainer : Kisung You  (kyoustat@gmail.com)")
+  packageStartupMessage("** Website    : https://kyoustat.com/Riemann/")
+  packageStartupMessage("**")
+  packageStartupMessage("** Please share any bugs or suggestions to the maintainer.")
+  packageStartupMessage("** ------------------------------------------------------- **")
+}
+
+.onUnload <- function(libpath) {
+  library.dynam.unload("Riemann", libpath)
+}