+++ NEWS.md
@@ -2,6 +2,7 @@
 
 * Interfaces to some of the functions are changed to emphasize key parameters only for users.
 * Added functionalities : `riem.scNJW()`, `riem.scSM()`, `riem.scUL()`, `riem.wasserstein()`.
+* Support for landmark-based shape shape with `wrap.landmark()` added. 
 
 # Riemann 0.1.0